Update access.expiration to be a float
As of spring SHA 4cc00bdc70d0da4fe52606421fe430375d18e376, the
expiration type for keys is a float. Update perfrunner to match.

+++ perfrunner/settings.py
@@ -426,7 +426,7 @@ class PhaseSettings(object):
         self.doc_gen = options.get('doc_gen', self.DOC_GEN)
         self.size = int(options.get('size', self.SIZE))
         self.items = int(options.get('items', self.ITEMS))
-        self.expiration = int(options.get('expiration', self.EXPIRATION))
+        self.expiration = float(options.get('expiration', self.EXPIRATION))
         self.working_set = float(options.get('working_set', self.WORKING_SET))
         self.working_set_access = int(options.get('working_set_access',
                                                   self.WORKING_SET_ACCESS))
